requirement.txt
streamlit
numpy
pandas
matplotlib
scikit-learn
Pillow